ROX-26719: Extract common TLS issuer code
