+++ sensor/kubernetes/certrefresh/localscanner_tls_issuer.go
@@ -23,23 +23,6 @@ import (
 var (
 	log = logging.LoggerForModule()
 
-	startTimeout                         = 6 * time.Minute
-	fetchSensorDeploymentOwnerRefBackoff = wait.Backoff{
-		Duration: 10 * time.Millisecond,
-		Factor:   3,
-		Jitter:   0.1,
-		Steps:    10,
-		Cap:      startTimeout,
-	}
-	processMessageTimeout = 5 * time.Second
-	certRefreshTimeout    = 5 * time.Minute
-	certRefreshBackoff    = wait.Backoff{
-		Duration: 5 * time.Second,
-		Factor:   3.0,
-		Jitter:   0.1,
-		Steps:    5,
-		Cap:      10 * time.Minute,
-	}
 	_ common.SensorComponent = (*localScannerTLSIssuerImpl)(nil)
 )
